tests: Move common codes
Move vklayertest.cpp into layer_validation_tests.cpp and change
vklayertest.h to layer_validation_tests.h
